Refactor: Split Warzone checks from FlagWarConfig
Warzones are a feature in their own right - for now, placing them in the 'com.palmergames.bukkit.towny.war.common' package as a 'WarZoneConfig' class.
