+++ src/com/palmergames/bukkit/towny/war/flagwar/FlagWarConfig.java
@@ -7,9 +7,6 @@ import com.palmergames.util.TimeTools;
 import org.bukkit.Material;
 import org.bukkit.Tag;
 
-import java.util.List;
-import java.util.Set;
-
 public class FlagWarConfig {
 
 	public static final Material[] woolColors = new Material[] {
@@ -20,8 +17,6 @@ public class FlagWarConfig {
 	private static Material flagBaseMaterial = null;
 	private static Material flagLightMaterial = null;
 	private static Material beaconWireFrameMaterial = null;
-
-	private static Set<Material> editableMaterialsInWarZone = null;
 
 	public static boolean isAffectedMaterial(Material material) {
 
@@ -116,51 +111,6 @@ public class FlagWarConfig {
 	public static int getMinPlayersOnlineInNationForWar() {
 
 		return TownySettings.getInt(ConfigNodes.WAR_ENEMY_MIN_PLAYERS_ONLINE_IN_NATION);
-	}
-
-	public static void setEditableMaterialsInWarZone(Set<Material> editableMaterialsInWarZone) {
-
-		FlagWarConfig.editableMaterialsInWarZone = editableMaterialsInWarZone;
-	}
-
-	public static boolean isEditableMaterialInWarZone(Material material) {
-
-		return FlagWarConfig.editableMaterialsInWarZone.contains(material);
-	}
-
-	public static boolean isAllowingSwitchesInWarZone() {
-
-		return TownySettings.getBoolean(ConfigNodes.WAR_WARZONE_SWITCH);
-	}
-
-	public static boolean isAllowingFireInWarZone() {
-
-		return TownySettings.getBoolean(ConfigNodes.WAR_WARZONE_FIRE);
-	}
-
-	public static boolean isAllowingItemUseInWarZone() {
-
-		return TownySettings.getBoolean(ConfigNodes.WAR_WARZONE_ITEM_USE);
-	}
-
-	public static boolean isAllowingExplosionsInWarZone() {
-
-		return TownySettings.getBoolean(ConfigNodes.WAR_WARZONE_EXPLOSIONS);
-	}
-
-	public static boolean explosionsBreakBlocksInWarZone() {
-
-		return TownySettings.getBoolean(ConfigNodes.WAR_WARZONE_EXPLOSIONS_BREAK_BLOCKS);
-	}
-
-	public static boolean regenBlocksAfterExplosionInWarZone() {
-
-		return TownySettings.getBoolean(ConfigNodes.WAR_WARZONE_EXPLOSIONS_REGEN_BLOCKS);
-	}
-	
-	public static List<String> getExplosionsIgnoreList() {
-
-		return TownySettings.getStrArr(ConfigNodes.WAR_WARZONE_EXPLOSIONS_IGNORE_LIST);
 	}
 
 	public static double getWonTownblockReward() {
